+++ packages/ui-buttons/src/index.tsx
@@ -10,5 +10,6 @@ export * from './ButtonPrimaryInvert';
 export * from './ButtonSecondary';
 export * from './ButtonSubmit';
 export * from './ButtonSubmitInvert';
+export * from './ButtonTertiary';
 
 export * from './ButtonText';
